CP-54351 move LUNperVDI from /opt
