+++ Makefile
@@ -42,6 +42,7 @@ SM_LIBS += fjournaler
 SM_LIBS += ipc
 SM_LIBS += journaler
 SM_LIBS += lock_queue
+SM_LIBS += LUNperVDI
 SM_LIBS += lvhdutil
 SM_LIBS += lvmanager
 SM_LIBS += lvmcache
